Tri-light setup
Placed some more lights into the scene to try and create the mood that I'm going for. Put 2 spotlights at the POV of the camera and to the right of the camera, and put a directional light to the right of the camera to act as the sun coming through a window. Had to disable the shadows on the directional light as the shadows it was generating were not pleasant