Camera movement with tween js
Added functionality that allows the camera position to be tweened with tween js. Made a separate file to handle the tweening functions, which each copy the current position of the camera and then tween it to a given target position, making sure the camera's position is updated to the current tween values.