Navigation buttons
Added another object array for three navigation buttons that will probably (but not guaranteed) replace the arrow keys in moving around the website. Implemented code that allows them to have a function that uses the camera animation function. Also reworked the multiple camera animation functions into a similar function since most the code was the same.