default nav button functionality
changed the navigation button object array so they can disable the html that they aren't telling the camera to move towards, and tell the camera to activate their respective html, allowing for a removal of one of the large if else statements in the tween function for the camera. Broke the arrow keys in the process but will work on those later.